requirements.txt
mysql-connector-python
PySimpleGui
